vlc-python-env/vlc-python/requirements.txt
aniso8601==3.0.2
Click==7.0
Flask==1.0.2
Flask-Jsonpify==1.5.0
Flask-RESTful==0.3.6
itsdangerous==0.24
Jinja2==2.10
MarkupSafe==1.0
python-vlc==3.0.102
pytz==2018.5
six==1.11.0
Werkzeug==0.14.1
